requirements.txt
absl-py==0.15.0
apache-beam==2.34.0
APScheduler==3.6.3
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
avro-python3==1.10.2
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
chart-studio==1.1.0
click==8.0.3
colorama==0.4.4
colorlover==0.3.0
contextlib2==21.6.0
crcmod==1.7
cryptography==36.0.0
cufflinks==0.17.3
cycler==0.11.0
Cython==0.29.25
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.1.1
Django==3.2.9
django-crispy-forms==1.13.0
dm-tree==0.1.6
docopt==0.6.2
docutils==0.18.1
entrypoints==0.3
fastavro==1.4.7
Flask==2.0.2
Flask-Compress==1.10.1
flatbuffers==1.12
fonttools==4.28.2
future==0.18.2
gast==0.4.0
gensim==4.1.2
gin-config==0.5.0
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.34.1
h5py==2.10.0
hdfs==2.6.0
httplib2==0.19.1
idna==3.3
imageai==2.1.6
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaggle==1.5.12
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
keyring==23.4.0
kiwisolver==1.3.2
libclang==12.0.0
lvis==0.5.3
lxml==4.6.4
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.2
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.5
notebook==6.4.6
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
object-detection==0.1
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
orjson==3.6.5
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.8.2
platformdirs==2.4.0
plotly==5.4.0
portalocker==2.3.2
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.23
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pycparser==2.21
pydot==1.4.2
Pygments==2.10.0
pymongo==3.12.2
pyparsing==2.4.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==5.0.2
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
readme-renderer==30.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rfc3986==1.5.0
rsa==4.8
sacrebleu==2.0.0
scikit-learn==1.0.1
scipy==1.4.1
seaborn==0.11.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
setuptools-scm==6.3.2
six==1.15.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.1
SQLAlchemy==1.4.27
sqlparse==0.4.2
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io==0.22.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-metadata==1.5.0
tensorflow-model-optimization==0.7.0
tensorflow-object-detection-api==0.1.1
tensorflow-text==2.7.3
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
textblob==0.17.1
tf-models-official==2.7.0
tf-slim==1.1.0
tflearn==0.5.0
threadpoolctl==3.0.0
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tweepy==4.4.0
twine==3.7.0
typeguard==2.13.2
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wordcloud==1.8.1
wrapt==1.12.1
zipp==3.6.0
